+++ src/lib.rs
@@ -109,18 +109,18 @@ struct DummyComponent;
 If your component contains an entity then it cannot be deserialized as is
 because entity IDs are different on server and client. The client should do the
 mapping. Therefore, to replicate such components properly, they need implement
-[`MapNetworkEntities`] and registered using [`AppReplicationExt::replicate_mapped()`]:
+[`MapEntities`] and registered using [`AppReplicationExt::replicate_mapped()`]:
 
 ```
-# use bevy::prelude::*;
+# use bevy::{prelude::*, ecs::entity::{EntityMapper, MapEntities}};
 # use bevy_replicon::prelude::*;
 # use serde::{Deserialize, Serialize};
 #[derive(Component, Deserialize, Serialize)]
 struct MappedComponent(Entity);
 
-impl MapNetworkEntities for MappedComponent {
-    fn map_entities<T: Mapper>(&mut self, mapper: &mut T) {
-        self.0 = mapper.map(self.0);
+impl MapEntities for MappedComponent {
+    fn map_entities<T: EntityMapper>(&mut self, mapper: &mut T) {
+        self.0 = mapper.map_entity(self.0);
     }
 }
 ```
@@ -299,7 +299,7 @@ struct DummyEvent;
 
 Just like components, if an event contains an entity, then the client should
 map it before sending it to the server.
-To do this, use [`ClientEventAppExt::add_mapped_client_event()`] and implement [`MapNetworkEntities`]:
+To do this, use [`ClientEventAppExt::add_mapped_client_event()`] and implement [`MapEntities`]:
 
 ```
 # use bevy::{prelude::*, ecs::entity::MapEntities};
@@ -482,9 +482,7 @@ pub mod prelude {
         renet::{RenetClient, RenetServer},
         replicon_core::{
             dont_replicate::{CommandDontReplicateExt, EntityDontReplicateExt},
-            replication_rules::{
-                AppReplicationExt, MapNetworkEntities, Mapper, Replication, ReplicationRules,
-            },
+            replication_rules::{AppReplicationExt, Replication, ReplicationRules},
             replicon_tick::RepliconTick,
             NetworkChannels, ReplicationChannel, RepliconCorePlugin,
         },