+++ src/network_event.rs
@@ -6,9 +6,6 @@ use std::{marker::PhantomData, time::Duration};
 use bevy::{ecs::entity::EntityHashMap, prelude::*};
 use bevy_renet::renet::SendType;
 
-#[allow(deprecated)]
-use crate::replicon_core::replication_rules::Mapper;
-
 /// Holds a server's channel ID for `T`.
 #[derive(Resource)]
 pub struct ServerEventChannel<T> {
@@ -102,16 +99,6 @@ impl From<EventType> for SendType {
 ///
 /// Panics if a mapping doesn't exists.
 pub struct EventMapper<'a>(pub &'a EntityHashMap<Entity>);
-
-#[allow(deprecated)]
-impl Mapper for EventMapper<'_> {
-    fn map(&mut self, entity: Entity) -> Entity {
-        *self
-            .0
-            .get(&entity)
-            .unwrap_or_else(|| panic!("entity {entity:?} should be mappable"))
-    }
-}
 
 impl EntityMapper for EventMapper<'_> {
     fn map_entity(&mut self, entity: Entity) -> Entity {