+++ tests/client_event.rs
@@ -115,13 +115,6 @@ struct DummyEvent;
 #[derive(Deserialize, Event, Serialize, Clone)]
 struct MappedEvent(Entity);
 
-#[allow(deprecated)]
-impl MapNetworkEntities for MappedEvent {
-    fn map_entities<T: Mapper>(&mut self, mapper: &mut T) {
-        self.0 = mapper.map(self.0);
-    }
-}
-
 impl MapEntities for MappedEvent {
     fn map_entities<M: EntityMapper>(&mut self, entity_mapper: &mut M) {
         self.0 = entity_mapper.map_entity(self.0);